requirements.txt
asgiref==3.6.0
Django==4.1.4
django-froala-editor==4.0.16
gunicorn==20.1.0
Pillow==9.3.0
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.3.0
